refactor: 레스토랑 목데이터 data 폴더 하위로 이동
